+++ tests/test_endtoend.py
@@ -385,6 +385,68 @@ def _get_swap_result(probe):
     return probe.result()
 
 
+@with_probe
+def test_resolve_names(probe):
+    """
+    Name resolution is performed in the context of the Kubernetes cluster.
+    """
+    result = probe.result()
+    result.write("gethostbyname {}.{}".format(
+        result.deployment_ident.name,
+        result.deployment_ident.namespace,
+    ))
+    success, reply = loads(result.read())
+    assert success and IPv4Address(reply), reply
+
+
+@with_probe
+def test_resolve_names_failure(probe):
+    """
+    Attempted resolution of non-existent names results in blah blah blah
+    """
+    result = probe.result()
+    result.write("gethostbyname example.invalid")
+    success, reply = loads(result.read())
+    assert not success, reply
+
+
+@with_probe
+def test_resolve_addresses(probe):
+    """
+    Reverse name resolution is performed in the context of the Kubernetes
+    cluster.
+    """
+    if probe.method.name == "inject-tcp":
+        pytest.xfail("Issue 546")
+
+    result = probe.result()
+    result.write("gethostbyaddr 4.2.2.1")
+    success, reply = loads(result.read())
+    assert success and "level3.net" in reply[0], reply
+
+
+@with_probe
+def test_resolve_addresses_failure(probe):
+    """
+    Attempted reverse name resolution for addresses with no corresponding
+    names results in blah blah blah.
+    """
+    result = probe.result()
+    # RFC 6890 - An address from TEST-NET-1.  Selected in hopes it will never
+    # reverse resolve to anything.
+    result.write("gethostbyaddr 192.0.2.1")
+    success, reply = loads(result.read())
+    assert (
+        # musl libc behaves like this - so container mode will produce this
+        # result.
+        (success and reply[0] == '192.0.2.1') or
+        # glibc behaves like this - other modes should produce this result
+        # (unless you happen to have musl libc installed for some other
+        # reason).
+        (not success)
+    ), (success, reply)
+
+
 @after_probe
 def test_swapdeployment_restores_container_image(probe):
     """